pyproject.toml
[tool.poetry]
name = "polyode"
version = "0.1.0"
description = ""
authors = ["EdwardDB"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
scipy = "^1.8.1"
jupyterlab = "^3.4.2"
matplotlib = "^3.5.2"
pytorch-lightning = "^1.6.4"
wandb = "^0.12.17"
ipdb = "^0.13.9"
torchdiffeq = "^0.2.3"
plotly = "^5.8.0"
pandas = "^1.4.2"
scikit-learn = "^1.1.1"
torchvision = "^0.12.0"
tqdm = "^4.64.0"
sktime = "^0.13.0"
tables = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
